Pull out kicker into own component
